src/app/requirements.txt
asgiref==3.3.4
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
# cryptography==3.4.7
defusedxml==0.7.1
Django==3.2.3
django-allauth==0.44.0
django-phone-auth==0.3
django-phonenumber-field==5.1.0
django-rest-auth==0.9.5
djangorestframework==3.12.4
gunicorn==20.1.0
idna==2.10
oauthlib==3.1.0
phonenumbers==8.12.24
Pillow==8.2.0
psycopg2-binary==2.8.6
pycparser==2.20
PyJWT==2.1.0
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
six==1.16.0
sqlparse==0.4.1
urllib3==1.26.5
